requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
python-multipart>=0.0.9

# OpenAI API client
openai>=1.0.0

# HTTP requests (for testing and external APIs)
requests>=2.31.0
beautifulsoup4>=4.12.0
chardet>=5.0.0  # For encoding detection

# Browser automation for JavaScript-heavy pages
playwright>=1.40.0

# Image processing / CV (no TensorFlow required)
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
# Optional OCR (disabled by default)
# pytesseract>=0.3.10
